Adding .h file for order_parser
Added .h file,
changed the import statement in "interaction_service.h" from "order_parser.cpp" to "order_parser.h".
Included core QT libraries in Cmake file and changed order_parser.cpp to implement definitions.
