+++ app/interaction_service/CMakeLists.txt
@@ -14,6 +14,7 @@ set(CMAKE_CXX_STANDARD_REQUIRED ON)
 find_package(QT NAMES Qt6 Qt5 COMPONENTS Widgets REQUIRED)
 find_package(Qt${QT_VERSION_MAJOR} COMPONENTS Widgets REQUIRED)
 find_package(Qt${QT_VERSION_MAJOR} COMPONENTS Network REQUIRED)
+find_package(Qt${QT_VERSION_MAJOR} COMPONENTS Core REQUIRED)
 
 
 set(PROJECT_SOURCES
@@ -26,6 +27,7 @@ set(PROJECT_SOURCES
         parser/pbf-parser.h
         interaction_service.cpp
         interaction_service.h
+        utils/order_parser.h
 )
 
 add_subdirectory(parser)
