+++ app/interaction_service/utils/order_parser.cpp
@@ -1,38 +1,22 @@
-#include <QFile>
-#include <QTextStream>
-#include <stdio.h>
-#include <QtDebug>
-#include <QMap>
-#include <QVector>
+#include "utils/order_parser.h"
 
-
-class Order{
-  public:
-    Order(){}
-    void set_id(int id){
-        this->id = id;
-        qDebug()<< "set id";
-    }
-    void set_location(QString location){
-        this->location = location;
-        qDebug()<<"set location";
-    }
-    void set_geolocation(QString geo){
-        this->geolocation = geo;
-        qDebug() << "set Geolocation";
-    }
-    void set_other(QString header,QString value){
-        // overloads the value if header already exists
-        this->other_dict.insert(header,value);
-        qDebug() << "set other to key,value:" << header << ", " << value;
-    }
-  private:
-    int id; //order ID
-    QString location; //order geolocation
-    QString geolocation;
-    QMap<QString , QString> other_dict;
-};
-
+void Order::set_id(int id){
+    this->id = id;
+    qDebug()<< "set id";
+}
+void Order::set_location(QString location){
+    this->location = location;
+    qDebug()<<"set location";
+}
+void Order::set_geolocation(QString geo){
+    this->geolocation = geo;
+    qDebug() << "set Geolocation";
+}
+void Order::set_other(QString header,QString value){
+    // overloads the value if header already exists
+    this->other_dict.insert(header,value);
+    qDebug() << "set other to key,value:" << header << ", " << value;
+}
 //vector<string> headers
 //split()
 //split_line[i] corresponds to headers[i]
